+++ CHANGELOG.md
@@ -1,8 +1,14 @@
-# 2.6.0
+# 2.6.0 (2026-xx-xx)
+
+FEATURES:
+
+- **New Data Source:** `netapp-ontap_s3_user` and `netapp-ontap_s3_users`([#507](https://github.com/NetApp/terraform-provider-netapp-ontap/issues/507))
+- **New Resource:** `netapp-ontap_s3_user` ([#508](https://github.com/NetApp/terraform-provider-netapp-ontap/issues/508))
 
 BUG FIXES:
 
-- **netapp-ontap_san_lun-maps**: Fixed Value Conversion Error when optional filter fields are not set([#651](https://github.com/NetApp/terraform-provider-netapp-ontap/issues/651))
+- **netapp-ontap_snapmirror_policy resource**: fixed issue with create when `retention` is not set in config ([#647](https://github.com/NetApp/terraform-provider-netapp-ontap/issues/647))
+- **netapp-ontap_san_lun-maps**: Fixed Value Conversion Error when optional filter fields are not set ([#651](https://github.com/NetApp/terraform-provider-netapp-ontap/issues/651))
 
 # 2.5.0 (2026-03-12)